modify : 디렉토리 아키텍쳐 수정
