+++ apps/user/src/app/(auth)/components/signup-sections/style.css.ts
