+++ apps/user/src/app/main/components/not-approved/style.css.ts
