+++ apps/user/src/app/my-page/components/stages/type.ts
