+++ apps/user/src/app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
-import NotApproved from "../shared/components/notApproved/NotApproved";
-import useUserStore from "../shared/stores/useUserStore";
+import useUserStore from "../shared/stores/use-user-store";
+import NotApproved from "./main/components/not-approved/not-approved";
 
 export default function Home() {
   const { user } = useUserStore();
